requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiolimiter==1.0.0
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.10.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pip==23.1.1
python-telegram-bot==20.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
setuptools==67.7.1
six==1.16.0
sniffio==1.3.0
tornado==6.2
tzdata==2023.3
tzlocal==4.3
wheel==0.40.0
yarl==1.8.2
